feat: comprehensive frontend refactoring - eliminate duplicates and reorganize structure
