--- apps/web/app/api/proposals/actions[dep].ts
@@ -1,35 +1,0 @@
-// "use server";
-
-// // This file is deprecated. Please import from the new location:
-// // import { createProposal, uploadProposalFile } from "@/features/proposals/api";
-
-// // Redirecting to new location to prevent breaking changes
-// import { redirect } from "next/navigation";
-
-// export async function createProposal() {
-//   console.error(
-//     "This file is deprecated. Please import from @/features/proposals/api"
-//   );
-//   redirect("/error?message=API+moved");
-// }
-
-// export async function uploadProposalFile() {
-//   console.error(
-//     "This file is deprecated. Please import from @/features/proposals/api"
-//   );
-//   redirect("/error?message=API+moved");
-// }
-
-// export async function uploadProposalFileEnhanced() {
-//   console.error(
-//     "This file is deprecated. Please import from @/features/proposals/api"
-//   );
-//   redirect("/error?message=API+moved");
-// }
-
-// export async function createProposalWithQuestions() {
-//   console.error(
-//     "This file is deprecated. Please import from @/features/proposals/api"
-//   );
-//   redirect("/error?message=API+moved");
-// }
