--- apps/web/src/components/auth/LoginForm.tsx
@@ -1,16 +1,0 @@
-/**
- * Login form component
- */
-
-export function LoginForm() {
-  return (
-    <div className="w-full max-w-md space-y-6">
-      <div className="text-center">
-        <h1 className="text-2xl font-bold">Login</h1>
-        <p className="text-muted-foreground">
-          Login functionality not implemented
-        </p>
-      </div>
-    </div>
-  );
-}
