+++ apps/web/src/lib/utils/diagnostic-tools.ts
