+++ apps/web/src/lib/utils/ensure-tool-responses.ts
